+++ pyproject.toml
@@ -8,10 +8,16 @@ authors = [
 readme = "README.md"
 requires-python = ">=3.12,<3.15"
 dependencies = [
-    "pyinstaller (>=6.15.0,<7.0.0)"
+    "pyinstaller (>=6.15.0,<7.0.0)",
+    "pyopengl (>=3.1.10,<4.0.0)",
+    "pyopengl-accelerate (>=3.1.10,<4.0.0)",
+    "databases (>=0.9.0,<0.10.0)",
+    "aiosqlite (>=0.21.0,<0.22.0)"
 ]
 packages = []
 
+[tool.pytest.ini_options]
+pythonpath = ["src"]
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
